Resources terminado
Resources implementado, junto de algumas mudanças necessárias, além de correção de erros nos calculos de tiles

+++ include/Sprite.hpp
@@ -21,7 +21,7 @@ class Sprite: public Component{//implementar update, is e render da classe mãe 
     void Open(std::string file);
     void SetClip(int x, int y, int w, int h);
     void  Render();
-    void  Render(float x, float y);
+    void  Render(int x, int y);
     int GetWidth();
     int GetHeight();
     bool IsOpen();
